requirements.txt
streamlit>=1.30,<2.0

numpy==1.26.4
pandas>=1.5,<3.0
scikit-learn>=1.2,<2.0

sentence-transformers>=2.2
transformers>=4.35
torch>=2.0
tqdm

google-genai
python-dotenv
nltk
kagglehub
